+++ packages/cindel/lib/cindel.dart
@@ -5,7 +5,12 @@ library cindel;
 export 'package:cindel_annotations/cindel_annotations.dart';
 
 export 'src/binary_document.dart'
-    show cindelDecodeBinaryDocument, cindelEncodeBinaryDocument;
+    show
+        CindelBinaryFieldType,
+        cindelDecodeBinaryDocument,
+        cindelDecodeSchemaBinaryDocument,
+        cindelEncodeBinaryDocument,
+        cindelEncodeSchemaBinaryDocument;
 export 'src/cindel.dart';
 export 'src/database.dart';
 export 'src/query.dart';